requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
hashids==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
psycopg2-binary==2.9.10
qrcode==8.0
setuptools==75.6.0
Werkzeug==3.1.3
